pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"


[project]
name = "semaine-nsi"
version = "0.1"
requires-python = ">=3.10"
dependencies = [
    "glfw ~= 2.10.0",
    "numpy ~= 2.3.3",
    "pyopengl ~= 3.1.10",
    "imgui-bundle ~= 1.92.4",
]